requirements.txt
fastapi==0.135.1
httpx==0.28.1
motor==3.7.1
passlib==1.7.4
pydantic==2.12.5
pymongo==4.16.0
python-dotenv==1.2.2
python_jose==3.5.0
